// src/modules/merchandising/reports/report/general-block-fabric-status-report/general-block-f-status-rpt-type.ts
export type GeneralBlockFabricStatusReportType = {
  COMPANY_NAME: string;
  COMPANY_ADDRESS: string;
  WORK_ORDER_NO: string;
  WORK_ORDER_ID: string;
  PI_NO: string;
  BUYER: string;
  STYLE_ID: number;
  STYLENO: string;
  STYLENAME: string;
  FABRIC_ID: number;
  FABRIC: string;
  FABRIC_COLOR: string;
  GMT_COLOR: string;
  WO_QTY: number;
  RCV_QTY: number;
  ALLOCATED_QTY: number;
  ALLOCATED_BALANCE_QTY: number;
  UOM: string;
  FABRIC_DIA: string;
  FABRIC_GSM: string;
  MATERIAL_DESCRIPTION: string;
  ORDER_REFERENCE: string;
  SEASON: string;
};